common controller

+++ app/Services/CategoryService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services;
 use App\Models\Category;
+use Illuminate\Support\Facades\DB;
 class CategoryService
 {
     public function getAll(){
@@ -18,13 +19,17 @@ class CategoryService
         return $category->update($validated);
     }
     public function delete($category){
-        return $category->delete();
+        if ($category) {
+            return DB::transaction(function () use ($category) {
+                $this->detach_with_products($category);
+                return $category->delete();
+            });
+        }
+
     }
     public function detach_with_products(Category $category){
         $products = $category->products()->get();
         if($products)
-        foreach ($products as $product) {
-            $category->products()->detach($product);
-        }
+            $category->products()->detach($products);
     }
 }
